+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "basic-blackjack",
-  "version": "0.2.9",
+  "version": "1.0.0",
   "private": true,
   "main": "./public/electron.js",
   "author": "devklick <devklick@gmail.com>",
