+++ src/App.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 import "./App.scss";
-import Table from "./components/Table/Table";
+import Table from "./components/Table";
 import { useGameSettingsStore } from "./stores/gameSettingsStore";
-import GameSettings from "./components/GameSettings/GameSettings";
-import CheckBox from "./components/CheckBox/CheckBox";
+import GameSettings from "./components/GameSettings";
+import CheckBox from "./components/CheckBox";
 
 const App: React.FC = () => {
   const { settingsModalOpen, setSettingsModelOpen } = useGameSettingsStore();
