+++ src/components/Buttons/index.ts
@@ -1,0 +1,3 @@
+import Buttons from "./Buttons";
+
+export default Buttons;
