+++ src/components/Card/index.ts
@@ -1,0 +1,7 @@
+import Card from "./Card";
+
+export * from "./card.types";
+export * from "./hooks/useCardPile";
+export * from "./hooks/useDeck";
+
+export default Card;
