+++ src/components/CardRow/index.ts
@@ -1,0 +1,3 @@
+import CardRow from "./CardRow";
+
+export default CardRow;
