+++ src/components/CheckBox/index.ts
@@ -1,0 +1,3 @@
+import CheckBox from "./CheckBox";
+
+export default CheckBox;
