+++ src/components/GameSettings/index.ts
@@ -1,0 +1,3 @@
+import GameSettings from "./GameSettings";
+
+export default GameSettings;
