+++ src/components/InfoHud/index.ts
@@ -1,0 +1,5 @@
+import InfoHud from "./InfoHud";
+
+export * from "./hooks/useGameStats";
+
+export default InfoHud;
