+++ src/components/ScoreBoard/index.ts
@@ -1,0 +1,5 @@
+import ScoreBoard from "./ScoreBoard";
+
+export * from "./ScoreBoard";
+
+export default ScoreBoard;
