+++ src/components/Table/index.ts
@@ -1,0 +1,5 @@
+import Table from "./Table";
+
+export * from "./hooks/useGame";
+
+export default Table;
