+++ src/components/YesNoPopUp/index.ts
@@ -1,0 +1,3 @@
+import YesNoPopUp from "./YesNoPopUp";
+
+export default YesNoPopUp;
